+++ packages/themes/neutral/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @xds/theme-neutral
 
+# 0.1.8
+
+---
+
 # 0.1.7
 
 #### Fixes